refactor(cli): reunite the announce record's two halves, drop the dial strategy
The consumer half of the signed-address work is finished here. `shard_cmd.rs`
now carries addresses and does nothing network-shaped with them.

`connect_chain_entry` is gone. It was a dial strategy — try each published
address, then fall back to a bare-PeerId dial — written at the one layer that
cannot make it stick, since a pre-connect only helps while its connection
lasts. Its three call sites (`cmd_shard_run`, `run_streaming_inner`,
`shard_api`) each become one best-effort `connect_peer_with_addrs`, which hands
the addresses to the daemon's learned map; the retry, the ordering and the DHT
fallback are the swarm's business now, and every later dial to the peer gets
them, not just the pre-connect. The one branch left in the CLI is a legacy
peer: `do_connect` rejects a request carrying no addresses as malformed, so an
empty `dial_addrs` — a binary predating `addrs_signed`, or a record that failed
verification — keeps the old `/p2p/<id>` dial.

`verified_dial_addrs` goes with it. It existed only to re-attach `/p2p/<peer>`
for `connect_peer`; the daemon re-attaches the destination itself, so addresses
stay bare from `verified_addrs` all the way to the swarm and
`BlockServerEntry::dial_addrs` is `Vec<Multiaddr>` rather than `Vec<String>`.
One conversion instead of two, and no place left that can produce
`/p2p/<id>/p2p/<id>`.

`ServerInfoFields` and the three `decode_server_info_*` functions move to
`announce.rs`, beside `DHTServerInfo::to_msgpack`. A separate `server_record.rs`
was the alternative and would have kept the split it exists to close: the
encoder and the decoder are one published wire format, every bug they have had
was a disagreement between them, and side by side the decoder can be tested
against the real producer instead of a hand-built map. `announce.rs` grows by
about a fifth, which is well short of unwieldy for this crate.

The decode tests move with the decoder, expectations updated to bare
addresses: the direct and circuit round trips through `to_msgpack`, the forged
signer, the dictionary subkey/value mismatch and its honest twin, legacy bytes
and an unparseable envelope. The legacy-signing-domain and address-cap cases
are dropped from here — `kwaai_p2p::peer_record` tests those against
`verified_addrs` directly, which is where they belong now.
